requirements.txt
aiogram>=3.4.1
asyncpg>=0.29.0
python-dotenv>=1.0.1
reportlab>=4.0.8
openpyxl>=3.1.2
yookassa>=2.3.5
pydantic>=2.6.0